requirements.txt
aiohttp
numpy
python-telegram-bot
python-dotenv
